GameTimer long verbouwd naar float
We verloren het comma getal wat ik in de vorige commit uitlegde

+++ Bermuda/Bermuda/GameTimer.cpp
@@ -40,7 +40,7 @@ void GameTimer::updateGameTime(double frameTime)
 	this->updateDay();
 }
 
-long GameTimer::getGameTime()
+float GameTimer::getGameTime()
 {
 	return this->gameTime;
 }
